cv-optimizer-be/requirements.txt
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.3
openai==0.27.0
PyPDF2==3.0.1
diff-match-patch
diff-match-patch-python
pillow
